Git-change update now amends commit on certain flags.
If at least one of --reviewers, --cc or --bug was passed to git-change
update, the HEAD commit is amended (regardless of whether there are
any staged files) in order to send a new patch set to Gerrit.

commit_change now takes git-commit args as a parameter and returns the
output from git-commit.