Fix --merge_commit bug when removing the HEAD commit.
The --merge_commit flag causes the HEAD (merge) commit of the tracking
branch to be removed with git-reset. Prior to this change it passed
--soft to git-reset which was wrong as it would leave files from the
commit in the index. This change switches to using --hard instead.

Since git reset --hard wipes out the index and working tree, this
change also adds a check for dirty files in either the index or
working tree and exits with an error if they're not clean.