Remove ASCII encoding.
This breaks when there are non-ASCII characters in the
commit message(or branch name, or anything else we get back from
gerrit), and subprocess can now accept unicode objects containing
non-ASCII characters as part of the command. This does mean that
this version of git-change will probably not work for versions of
Python < 2.7.